refactor: Use Hexagonal Package Structure
