--- src/main/java/com/hexagonal/courseregistration/user/AllUser.java
@@ -1,7 +1,0 @@
-package com.hexagonal.courseregistration.user;
-
-interface AllUser {
-  boolean exist(String idNumber, Authority authority);
-
-  void register(User user);
-}
