--- src/main/java/com/hexagonal/courseregistration/user/Authority.java
@@ -1,5 +1,0 @@
-package com.hexagonal.courseregistration.user;
-
-enum Authority {
-  STUDENT, PROFESSOR
-}
