--- src/main/java/com/hexagonal/courseregistration/user/JsonRegisterRequest.java
@@ -1,11 +1,0 @@
-package com.hexagonal.courseregistration.user;
-
-record JsonRegisterRequest(
-  String name,
-  String idNumber,
-  Authority authority
-) {
-  RegisterRequest toRegisterRequest() {
-    return new RegisterRequest(name, idNumber, authority);
-  }
-}
