--- src/main/java/com/hexagonal/courseregistration/user/RegisterRequest.java
@@ -1,8 +1,0 @@
-package com.hexagonal.courseregistration.user;
-
-public record RegisterRequest(
-  String name,
-  String idNumber,
-  Authority authority
-) {
-}
