+++ src/main/java/com/hexagonal/courseregistration/user/application/Registrant.java
@@ -1,8 +1,11 @@
-package com.hexagonal.courseregistration.user;
+package com.hexagonal.courseregistration.user.application;
 
+import com.hexagonal.courseregistration.user.application.port.out.AllUser;
+import com.hexagonal.courseregistration.user.domain.User;
 import lombok.RequiredArgsConstructor;
 
 @RequiredArgsConstructor
+public
 class Registrant {
   private final AllUser allUser;
