+++ src/main/java/com/hexagonal/courseregistration/user/domain/Authority.java
@@ -1,0 +1,5 @@
+package com.hexagonal.courseregistration.user.domain;
+
+public enum Authority {
+  STUDENT, PROFESSOR
+}
